..F....... [ZBX-8817] unified most of web scenario add/update/delete and other operation and audit messages where now instead of "Scenario...", "Web scenario..." is used. ..F....... [ZBX-8817] re-factored code for web scenario enable/disable and delete operations in frontend, so that required data is read and updated in bulk.
git-svn-id: svn://svn.zabbix.com/branches/dev@49803 f4e7fd02-d210-4e34-a1e0-d4093fb66cc2

+++ ZBX-8817/frontends/php/httpconf.php
@@ -122,35 +122,40 @@ if (isset($_REQUEST['new_httpstep'])) {
 	unset($_REQUEST['new_httpstep']);
 }
 
-if (isset($_REQUEST['delete']) && isset($_REQUEST['httptestid'])) {
+if (hasRequest('delete') && hasRequest('httptestid')) {
 	$result = false;
 
-	DBstart();
+	$httpTestId = getRequest('httptestid');
 
-	$host = DBfetch(DBselect(
-		'SELECT h.host'.
-		' FROM hosts h,httptest ht'.
-		' WHERE ht.hostid=h.hostid'.
-			' AND ht.httptestid='.zbx_dbstr($_REQUEST['httptestid'])
+	$httpTests = API::HttpTest()->get(array(
+		'output' => array('name'),
+		'httptestids' => array($httpTestId),
+		'selectHosts' => array('name'),
+		'editable' => true
 	));
 
-	$httptestData = get_httptest_by_httptestid($_REQUEST['httptestid']);
-	if ($httptestData) {
-		$result = API::HttpTest()->delete(array(getRequest('httptestid')));
+	if ($httpTests) {
+		DBstart();
+
+		$result = API::HttpTest()->delete(array($httpTestId));
 
 		if ($result) {
+			$httpTest = reset($httpTests);
+			$host = reset($httpTest['hosts']);
+
 			add_audit(AUDIT_ACTION_DELETE, AUDIT_RESOURCE_SCENARIO,
-				'Web scenario ['.$httptestData['name'].'] ['.$_REQUEST['httptestid'].'] Host ['.$host['host'].']'
+				_('Web scenario').' ['.$httpTest['name'].'] ['.$httpTestId.'] '._('Host').' ['.$host['name'].']'
 			);
+		}
+
+		$result = DBend($result);
+
+		if ($result) {
+			uncheckTableRows(getRequest('hostid'));
 		}
 	}
 	unset($_REQUEST['httptestid'], $_REQUEST['form']);
 
-	$result = DBend($result);
-
-	if ($result) {
-		uncheckTableRows(getRequest('hostid'));
-	}
 	show_messages($result, _('Web scenario deleted'), _('Cannot delete web scenario'));
 }
 elseif (isset($_REQUEST['clone']) && isset($_REQUEST['httptestid'])) {
@@ -195,12 +200,12 @@ elseif (hasRequest('add') || hasRequest('update')) {
 
 	if (hasRequest('update')) {
 		$action = AUDIT_ACTION_UPDATE;
-		$messageTrue = _('Scenario updated');
+		$messageTrue = _('Web scenario updated');
 		$messageFalse = _('Cannot update web scenario');
 	}
 	else {
 		$action = AUDIT_ACTION_ADD;
-		$messageTrue = _('Scenario added');
+		$messageTrue = _('Web scenario added');
 		$messageFalse = _('Cannot add web scenario');
 	}
 
@@ -317,7 +322,9 @@ elseif (hasRequest('add') || hasRequest('update')) {
 		}
 
 		$host = get_host_by_hostid($_REQUEST['hostid']);
-		add_audit($action, AUDIT_RESOURCE_SCENARIO, 'Scenario ['.$_REQUEST['name'].'] ['.$httpTestId.'] Host ['.$host['host'].']');
+		add_audit($action, AUDIT_RESOURCE_SCENARIO,
+			_('Web scenario').' ['.getRequest('name').'] ['.$httpTestId.'] '._('Host').' ['.$host['name'].']'
+		);
 
 		unset($_REQUEST['httptestid'], $_REQUEST['form']);
 		show_messages(true, $messageTrue);
@@ -333,37 +340,56 @@ elseif (hasRequest('add') || hasRequest('update')) {
 		show_messages(false, null, $messageFalse);
 	}
 }
-elseif (hasRequest('action') && str_in_array(getRequest('action'), array('httptest.massenable', 'httptest.massdisable')) && hasRequest('group_httptestid')) {
+elseif (hasRequest('action') && str_in_array(getRequest('action'), array('httptest.massenable', 'httptest.massdisable'))
+		&& hasRequest('group_httptestid') && is_array(getRequest('group_httptestid'))) {
 	$result = true;
-	$groupHttpTestId = getRequest('group_httptestid');
-	$enable = (getRequest('action') == 'httptest.massenable');
+	$httpTestIds = getRequest('group_httptestid');
+	$enable = (getRequest('action') === 'httptest.massenable');
 	$status = $enable ? HTTPTEST_STATUS_ACTIVE : HTTPTEST_STATUS_DISABLED;
 	$statusName = $enable ? 'enabled' : 'disabled';
 	$auditAction = $enable ? AUDIT_ACTION_ENABLE : AUDIT_ACTION_DISABLE;
 	$updated = 0;
 
-	DBStart();
+	$httpTests = API::HttpTest()->get(array(
+		'output' => array('httptestid', 'name', 'status'),
+		'httptestids' => $httpTestIds,
+		'selectHosts' => array('name'),
+		'editable' => true
+	));
 
-	foreach ($groupHttpTestId as $id) {
-		if (!($httpTestData = get_httptest_by_httptestid($id))) {
-			continue;
+	if ($httpTests) {
+		$httpTestsToUpdate = array();
+
+		DBstart();
+
+		foreach ($httpTests as $httpTest) {
+			// change status if it's necessary
+			if ($httpTest['status'] != $status) {
+				$httpTestsToUpdate[] = array(
+					'httptestid' => $httpTest['httptestid'],
+					'status' => $status
+				);
+			}
 		}
 
-		$result &= (bool) API::HttpTest()->update(array(
-			'httptestid' => $id,
-			'status' => $status
-		));
+		if ($httpTestsToUpdate) {
+			$result = API::HttpTest()->update($httpTestsToUpdate);
 
-		if ($result) {
-			$host = DBfetch(DBselect(
-				'SELECT h.host FROM hosts h,httptest ht WHERE ht.hostid=h.hostid AND ht.httptestid='.zbx_dbstr($id)
-			));
-			add_audit($auditAction, AUDIT_RESOURCE_SCENARIO,
-				'Scenario ['.$httpTestData['name'].'] ['.$id.'] Host ['.$host['host'].'] '.$statusName
-			);
+			if ($result) {
+				foreach ($httpTests as $httpTest) {
+					$host = reset($httpTest['hosts']);
+
+					add_audit($auditAction, AUDIT_RESOURCE_SCENARIO,
+						_('Web scenario').' ['.$httpTest['name'].'] ['.$httpTest['httptestid'].'] '.
+							_('Host').' ['.$host['name'].'] '.$statusName
+					);
+				}
+			}
 		}
 
-		$updated++;
+		$result = DBend($result);
+
+		$updated = count($httpTests);
 	}
 
 	$messageSuccess = $enable
@@ -373,17 +399,14 @@ elseif (hasRequest('action') && str_in_array(getRequest('action'), array('httpte
 		? _n('Cannot enable web scenario', 'Cannot enable web scenarios', $updated)
 		: _n('Cannot disable web scenario', 'Cannot disable web scenarios', $updated);
 
-	$result = DBend($result);
-
 	if ($result) {
 		uncheckTableRows(getRequest('hostid'));
 	}
+
 	show_messages($result, $messageSuccess, $messageFailed);
 }
-elseif (hasRequest('action')
-		&& getRequest('action') === 'httptest.massclearhistory'
-		&& hasRequest('group_httptestid')
-		&& is_array(getRequest('group_httptestid'))) {
+elseif (hasRequest('action') && getRequest('action') === 'httptest.massclearhistory'
+		&& hasRequest('group_httptestid') && is_array(getRequest('group_httptestid'))) {
 	$result = false;
 
 	$httpTestIds = getRequest('group_httptestid');
@@ -423,7 +446,8 @@ elseif (hasRequest('action')
 
 	show_messages($result, _('History cleared'), _('Cannot clear history'));
 }
-elseif (hasRequest('action') && getRequest('action') == 'httptest.massdelete' && hasRequest('group_httptestid')) {
+elseif (hasRequest('action') && getRequest('action') === 'httptest.massdelete'
+		&& hasRequest('group_httptestid') && is_array(getRequest('group_httptestid'))) {
 	$result = API::HttpTest()->delete(getRequest('group_httptestid'));
 
 	if ($result) {
